Abstract button icon styles
